% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_all}
\alias{run_all}
\title{Print two provided words.}
\usage{
run_all(word1, word2, word3)
}
\arguments{
\item{word1}{string provided by the user}

\item{word2}{string provided by the user}
}
\description{
Test method 2
}
