% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_two_words}
\alias{run_two_words}
\title{Print two provided words.}
\usage{
run_two_words(word1, word2)
}
\arguments{
\item{word1}{string provided by the user}

\item{word2}{string provided by the user}
}
\description{
Test method 2
}
